Fix attachments
