+++ imports/api/Api.js
@@ -2,21 +2,14 @@ import { Meteor } from "meteor/meteor";
 import Vue from "vue";
 
 export default {
-  call(...params) {
-    return new Promise((resolve, reject) => {
-      const methodName = params.shift();
-      Meteor.apply(
-        methodName,
-        params,
-        null,
-        (error, datas) => {
-          if (error) {
-            Vue.prototype.$notifyError(error);
-            reject(error);
-          }
-          resolve(datas);
-        }
-      );
-    });
+  async call(...params) {
+    const methodName = params.shift();
+    try {
+      const datas = await Meteor.applyAsync(methodName, params);
+      return datas;
+    } catch (error) {
+      Vue.prototype.$notifyError(error);
+      throw error;
+    }
   }
 };
