+++ imports/api/attachments/attachments.js
@@ -5,29 +5,14 @@ import {
   checkCanWriteTask,
   runAsUser
 } from "/imports/api/permissions/permissions";
+import { Log } from "meteor/logging";
 
 export const Attachments = new FilesCollection({
   collectionName: "Attachments",
   storagePath: Meteor.settings.attachmentsPath || "assets/app/uploads",
   allowClientCode: true, // Disallow remove files from Client
   onBeforeUpload(fileData) {
-    return runAsUser(this.userId, async function() {
-      if (fileData.meta?.taskId) {
-        try {
-          await checkCanWriteTask(fileData.meta.taskId);
-        } catch (error) {
-          return false;
-        }
-      }
-      if (fileData.meta?.projectId) {
-        try {
-          await checkCanWriteProject(fileData.meta.projectId);
-        } catch (error) {
-          return false;
-        }
-      }
-      return true;
-    });
+    return true;
   },
   onAfterUpload(fileRef) {
     if (Meteor.isServer) {
